+++ src/scimilarity/utils.py
@@ -89,6 +89,10 @@ def align_dataset(
             f"Dataset incompatible. Gene overlap of {gene_overlap} less than {gene_overlap_threshold}. Check that var.index uses gene symbols."
         )
 
+    # check if X is None, empty csr_matrix if so
+    if data.X is None:
+        data.X = csr_matrix(data.layers["counts"].shape)
+
     # check if X is dense, convert to csr_matrix if so
     if isinstance(data.X, np.ndarray):
         data.X = csr_matrix(data.X)
@@ -186,6 +190,53 @@ def filter_cells(
     return data
 
 
+def convert_id2symbol(
+    adata: "anndata.AnnData",
+    mapping_table: str,
+):
+    """Convert EnsembleIDs to gene symbols via a mapping table.
+
+    Parameters
+    ----------
+    adata: anndata.AnnData
+        Annotated data matrix with rows for cells and columns for genes.
+    mapping_table: str
+        A tsv file in Ensemble format with gene names in column "Gene stable ID".
+
+    Returns
+    -------
+    anndata.AnnData
+        AnnData object with gene symbols as the adata.var.index.
+
+    Examples
+    --------
+    >>> adata = consolidate_duplicate_symbols(adata)
+    """
+
+    import pandas as pd
+
+    id2name = pd.read_csv(mapping_table, delimiter="\t", index_col=0, dtype="unicode")[
+        "Gene name"
+    ]
+    id2name = (
+        id2name.reset_index().drop_duplicates().set_index("Gene stable ID")["Gene name"]
+    )
+    id2name = pd.concat(
+        [id2name, pd.Series(id2name.values, index=id2name.values).drop_duplicates()]
+    )
+
+    if not any(adata.var.index.isin(id2name.keys())) and "symbol" in adata.var.columns:
+        adata.var = adata.var.set_index("symbol", drop=False)
+    adata.var.index = adata.var.index.str.replace("'", "")
+    adata.var.index = adata.var.index.str.replace('"', "")
+    adata = adata[
+        :, (adata.var.index.isin(id2name.keys())) & ~(adata.var.index.isnull())
+    ].copy()
+    adata.var.index = id2name[adata.var.index]
+    adata.var.index.name = "symbol"
+    return adata
+
+
 def consolidate_duplicate_symbols(
     adata: "anndata.AnnData",
 ) -> "anndata.AnnData":
@@ -214,6 +265,8 @@ def consolidate_duplicate_symbols(
 
     if "counts" not in adata.layers:
         raise ValueError(f"Raw counts matrix not found in layers['counts'].")
+    if adata.X is None:
+        adata.X = csr_matrix(adata.layers["counts"].shape)
 
     gene_count = Counter(adata.var.index.values)
     dup_genes = {k for k in gene_count if gene_count[k] > 1}
@@ -249,7 +302,7 @@ def consolidate_duplicate_symbols(
 
 
 def get_centroid(
-    counts: Union["scipy.sparse.csr_matrix", "numpy.ndarray"]
+    counts: Union["scipy.sparse.csr_matrix", "numpy.ndarray"],
 ) -> "numpy.ndarray":
     """Get the centroid for a raw counts matrix.
 
@@ -551,9 +604,9 @@ def adata_from_tiledb(
     GENEURI: str = "gene_annotation",
     CELLURI: str = "cell_metadata",
     COUNTSURI: str = "counts",
-    config: Optional["tiledb.ctx.Config"] = None,
     lognorm: bool = True,
     target_sum: float = 1e4,
+    config: Optional["tiledb.ctx.Config"] = None,
 ):
     """Constructs an AnnData object from cells in tiledb.
 
@@ -565,20 +618,20 @@ def adata_from_tiledb(
         Base path of tiledb store
     gene_order: List[str], optional, default: None
         Gene order
-    SAMPLEURI: str, default:"sample_metadata"
+    SAMPLEURI: str, default: "sample_metadata"
         Sub path of sample metadata store
-    GENEURI: str, default:"gene_annotation"
+    GENEURI: str, default: "gene_annotation"
         Sub path of gene annotation store
-    CELLURI: str, default:"cell_metadata"
+    CELLURI: str, default: "cell_metadata"
         Sub path of cell metadata store
-    COUNTSURI: str, default:"counts"
+    COUNTSURI: str, default: "counts"
         Sub path of count matrix store
-    config: tiledb.ctx.Config, optional, default: None
-        Custom tiledb config
     lognorm: bool, default: True
         Whether to return log normalized expression instead of raw counts.
     target_sum: float, default: 1e4
         Target sum for log normalization.
+    config: tiledb.ctx.Config, optional, default: None
+        Custom tiledb config
 
     Returns
     -------
@@ -598,8 +651,7 @@ def adata_from_tiledb(
     import tiledb
 
     if config is None:
-        cfg = tiledb.Config()
-        cfg["sm.mem.total_budget"] = 50000000000  # 50G
+        cfg = tiledb.Config({"sm.mem.total_budget": 50000000000})  # 50G
     else:
         cfg = config